Provenance: added check by $ref attribute when used .
There are serveral options to enforce the correct use of the provenance attribute on dataset, table and/or row level. Choosen is to use the ./meta/provenance.json as the definition enforcer. The schema.json and row-meta-schema.json where adjusted to reference the ./meta/provenance.json. In the schema.json are general properties definied for the dataset and table. Addtional general properties on row level is definied in the row-meta-schema.json file. So both needed to be adjusted.