BBGA schema. [DS-559]
Though spaces in `id` names supposedly are not supported, and one should
use camelCase names instead, the schema-tools validator claims the
opposite. And as schema-tools' `schema create` sub commands happily and
correctly processes these `id` names with spaces, we have used exactly
that.

The sources for the tables `kerncijfers` and `statisieken` have a column
`variabelen` That column has been modelled as the relation to the parent
table `indicatoren definities`. In the DSO-API (database) these
relations will end up as a column called `indicator_definitie_id` in
both tables. It is not the most logical name for a column linking
variable values their definitions. But given the current limitation of
`schema-tools` it is what we have to work with.