[DS-549] Added bikeshed as a dependency.
Added version 2.3.3 of Bikeshed specifically and not version 2.3.0!
The latter has issue with rendering highlighted inline code.

Bikeshed initially looked very promissing. It promissed generating of
ReSpec like documentation with the convenience of Markdown. However it
uses its own Markdown parser that, although it claims conformance with
the Common Markdown specification, clearly is not conformant. In addition
to the earlier mentioned issue with inline code, it also has issues with
nested list that contain code blocks. Regardless of using version 2.3.1
or 2.3.0. Hence I am hesitant if this is the way forward for documenting
our spec.

Upgraded dev requirements as well. Most notably to get better error
message when using `pip-compile`.