[DS-549] Add Bikeshed/ReSpec version of AMS Schema Spec.
This is just a plain conversion of the Markdown document previously
posted on the Wiki with no added content. Although Bikeshed's major
feature, allowing Markdown content to generate ReSpec documents, is
promissing it falls short in practise. Bikeshed's custom Markdown parser
is flawed and I had had to tweak the document using a lot of trial and
error for it to render properly.

It has issues with:
- Rendering of highlighted inline code (had to downgrade Bikeshed).
- Code blocks in nested lists.
- Nested lists in general. For the third level to be picked up,
  lists need to be indented using 4 instead of 2 spaces.

Hence I'm not sure if Bikeshed is what we should be using.