Several fixes to the metaschema
Added new table fields that were not yet in the meta-schema:
- additionalRelations
- isTemporal

Relaxing pattern for identifiers (we are using camelCase now).

Made table "identifier" "array" type possible.

Remove "id" from the "required" fields of a table.

Changed "additionalProperties" from true -> false for the row schema.
This means that all possible attributes for row schemas have to be
made explicit in the metaschema.

Add more metaschema for the row definitions

Because we are using additionalProperties=false now
our metaschema should contain all the field attributes that are in use.

Furthermore:

For local schemas, the schema urls were not modified during publishing.
Effectively making them not useable for schema validation.
This has now been fixed by first creating a copy of the relevant files
 (metaschema and schemas) and modifying the schema urls before
 publishing (to a custom location).