Simplify singular FK relations
Singular FK relations (just a single identifier) were defined
as objects with one item (to mimick the ndjson response from GOB).

Has now been simplified to a simple scalar, using provenance
(jsonpath-based) to fetch the correct values from the ndjson.

The underlying goal is to remove the extra columns that are
automatically added for object-type relations in the schemas.