AB#5020 Prep Dockerfile for publishing from filesystem.
Publishing the schemas previously relied on downloading a ZIP file from
the Github repo. As we want to capture a specific moment in time we now
contain the schema files into the Docker build. This allows us to publish a
known set, regardless of the state of the Github repo (that might have
changed in the mean time) currently is in.

This required moving the Dockerfile one level up as we can't do `COPY
..`; `COPY` is always relative to the build == Dockerfile.