AB#5020 WIP Turn everything into a proper Python package.
In order to be able to publish specific revisions of the schemas in this
repo we could have mucked around with a Docker build that captures the
state of the repo into a specific build. Or we could choose to capture
the state of the repo into a versioned Python package. We chose the
latter. This involved the following steps:

- specifying setup.cfg and pyproject.toml files
- added tbump
- ditching requirements.* (they are now in setup.cfg)
- moving publish.py into a Python package with same name as this repo
- bundling schemas with the resulting Python package

The latter proved to be rather challening as the schemas live outside of
the Python package (eg they are not in a directory with a __init__.py
file). Moving them all into the newly created `amsterdam_schema` package
(under the `src` directory) would probably have confused other users of
this repo that expect the schemas to be in the root of the repo. Hence
that was not an option.

Initially it seemed that with a proper `MANIFEST.in` file and
`include_package_data = True` in `setup.cfg` the schemas (and datasets)
could be included. The `sdist` had everything we had specified in it.
But the `wheel` did not. Not building a wheel these days is not really
an option. Hence we had to look for a different option.

In the old days we could have used the `data_files` directive, though
that is currently being frowned upon in modern Python packaging. Still
we considered that option. The problem with `data_files` is that each
file needs to be specified separately as wildcards are not supported.
That is a problem in `setup.cfg` only distributions, as we can't use
Python code to dynamically build up the `data_files` data structure. So
we ditched that idea as well.

In hindsight the solution was simple. Git supports symbolic links and
Python packaging properly processes them. Hence we included symlinks to
the top level `datasets` and `schemas@*` directories from within the
`amsterdam_schema` package. Now we have the best of both worlds:

- data modelling users still find the schemas in a familiar location
 (root dir of the repo)
- and us, developers, that are responsible of publishing these schema's
  using Python scripts, can properly package these datasets and schemas
  together with the publishing code.

The Docker build is now rather simple. Instead of having knowledge where
the `publish` script expects the datasets and schemas, we can simply
build the Python package and install it. The Python package (the publish
in it) knows where everything is regardless of what the Docker build
did.

While we are at making this a proper Python package we have turned up
the linting and static type checking a couple of notches. This to bring
it more in line with what we are doing for the `schema-tools` package
(and hopefully others)

The `--use-local` option to the `publish` command now works quite
differently. Instead of working directly with the file system from a
cloned repo, it works with package resources. It the primary reason for
this whole packaging excercise.

NOTE regarding not havining `twine` included: We don't expect this
package to be published to PyPi. The package is only used in internal
builds.

The name mangling is weird. We should ditch it.