index.json on objectstore ((acc.)schemas.data.amsterdam.nl)
Schema definitions are stored at ((acc.)schemas.data.amsterdam.nl) following the structure:

/folder/schema_file

Most of the time the folder name and schema file name are the same. I.e. /wior/wior

Grex is an example where the structure is as follows:

/grex/projecten

In the index.json, which acts like a referrer for the schema file location, is a dictionary (this logic was recently changed in the context of schema versioning support). Containing as a key the schema file name. Which is great for situations like /wior/wior. Where also the dataset id corresponds to wior. For grex the folder name is identical to the datset id. In this case grex. The schema file name is the odd duck.

To keep the current logic and preventing issues the folder name is now used as key in the index.json. This should cope with grex and all other structures.