Extend HR dataset with new tables
Added 3 new tables:

- locaties
- vestigigingen
- sbiactiviteiten

These tables have been fixed for compatibility with the ndjson importer.

A new datatype has been introduced in the gob-model: IncompleteDate
In this datatype one of the date components (day/month/year)
is possibly missing. In that case `00` or `0000` is substituted.

This means that these incompleted dates cannot
be imported as dates in postgresql, so they are imported as strings,
and the `"format": "data"` is removed from the amsterdam schema
definition.