Add `cijfers` table to BBGA dataset.
This defines a convenience table `cijfers` that is the combination of
`kerncijfers` and `statistieken`. This was created for Jasper Swart as
he needed to be able to retrieve data from both tables in one API call.
With our current API and because of the different "dimensions" of these two
tables this could not be done otherwise.

With "dimensions" I mean that `kerncijfers` deals with values for a
given indicator for a given year for a given `gebiedcode15`.  However
`statistieken` deals with values for a given indicator and year, but NOT
for `gebiedcode15`. Eg it has one dimension less. This means that in the
tabel `cijfers` for a given indicator and year all `gebiedcode15` will
have the same statistical data. For the simple reason that the
statistical data is not for the `gebiedcode15` specifically, but for the
city as a whole.

Logically this table should not be here. It does not really make sense
to combine `kerncijfers` with `statistieken` when their dimensions do
not match up. However due to limitations in the current frontend
dashboard that queries this data this was deemed a necessity.