Add SCHEMA_URL env. var to github actions
When validating schemas, related schemas need to be loaded.
However, to find these schema's,
the SCHEMA_URL is needed by `schematools`.

This approach is sub-optimal, because the url is pointing
to the production version of the schema.
But, atm, it's all we have.

We need to come up with a better mechanism in the future.