Bugfixes for standvastgoed schema
The Amsterdam Schemas temporal fields can break the DSO-API if it
contains an uppercase or underscore in the field name, and if it has a
shortname, c.f. AB#24665.

The schema permissions apply call executes a query that includes an
uppercase if the schema dataset id contains an uppercase. However, all
table names are in snake_case. This means that the query that is
executed does not set the correct permissions if the dataset id is
standVastgoed, but does when it is standvastgoed, c.f. AB#24662