AB#25433 Sychronize BenK schemas
The amsterdam schemas for BenK/GOB datasets are not in sync
with the GOB model (json definition that GOB is using for its software).

The amsterdams schemas are created in the past with a tool provided by
GOB. However, in the meantime, the GOB model has evolved.

This PR does bring the amsterdam schema's in line with the current GOB
model.