AB#28259 Add publishing to an Azure storage account
The `publish` cli command now has te ability to write
the schema files to an Azure storage account.

An extra flag `--storage-type` has been added to the command
that has two possible values `swift` (default) and `azure`.

For the (current) swift option, the schemas for ACC and PRD are stored
on the same objectstore. For that reason,  there is a `--dp-env`
flag that can point to ACC or PRD.

For the azure option, the ACC and PRD storage accounts
need to be separate. So, on using the cli, this `--dp-env` should be
empty.

The `--container-prefix` is for the swift option combined
with the `--dp-env` to create the container name.

For azure, this `--container-prefix` should be the full name of the
container. So, the name of this option can be a bit misleading,
it should be changed when the migration to Azure is finished.