Add the v1.2.0 metaschema as package data
The publishing of metaschemas is only done if the folder with the
metaschema is part of the so-called `package_data`.

And `package_data` needs to be inside a package in the package_dir,
in our case `src/amsterdam_schema`.

So, a symlink needs to be place in that location, to make
the v1.2.0 metaschema publishable.