Allow datasets to use older minor schema versions
Replaces the constant in schema->$ref with a pattern match
that allows a dataset to use a different minor version than
that being used by the validator.
So long as a dataset is compatible with the new version in
every other respect the lag in minor version is allowed.
(minor versions should be backwards compatible anyway)