Add faker attribute to field meta schema.
The name of the attribute is still under consideration.
Could be that we want to change it to `format`.
However, that is not compatible with our current django model
generation yet.

Furthermore the regex for the $ref pattern for schema definitions
has been made a little more permissive, for easier local testing.