Introduce `"format": "json"` for brk and brp.
This (virtual) datasets have a level of object nested that is too deep,
it is not supported by the Amsterdam schema.

When those object fields are of format=json, this problem does not
occur. And, because these are proxy-ing dataset (no underlying tables)
anyway, it should be safe to change this.

Furthermore, some `reasonNonPublic` attributes needed to be added,
because validation on those has become more strict.