Rename schema v2.0.0 to v2 (#634)
Keeping around minor and patch versions of the metaschemas is an excuse
for not fixing issues with schemas. It also makes it much harder to use
git/GitHub because new versions are new files. Let's just keep a v2 file
from now on and edit that.

No schemas need to be changed, because none were explicitly referring to
v2.